refactor: Remove directives/ & interceptors/
