+++ src/app/converter/converter.component.spec.ts
@@ -2,9 +2,9 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ConverterComponent } from './converter.component';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ExchangeRateMockInterceptor } from 'src/app/interceptors';
 import { ExchangeRateService } from 'src/app/services';
 import { iso4217 } from 'src/app/exchange-rate/iso4217.enum';
+import { ExchangeRateMockInterceptor } from '../exchange-rate/exchange-rate-mock.interceptor';
 
 describe('ConverterComponent', () => {
     let component: ConverterComponent;
