--- src/app/directives/index.ts
@@ -1,1 +1,0 @@
-export * from './grid.directive';
