--- src/app/interceptors/index.ts
@@ -1,1 +1,0 @@
-export * from './exchange-rate-mock.interceptor';
